requirements.txt
asgiref==3.2.7
autopep8==1.5.2
beautifulsoup4==4.9.0
certifi==2020.4.5.1
chardet==3.0.4
defusedxml==0.6.0
Django==3.0.5
django-allauth==0.41.0
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==1.1.1
idna==2.9
joblib==0.15.1
numpy==1.18.4
oauthlib==3.1.0
pandas==1.0.3
pycodestyle==2.5.0
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
scikit-learn==0.23.0
scipy==1.4.1
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
threadpoolctl==2.0.0
urllib3==1.25.9
